force `webDialog()` calls to run on the main thread
The webDialog() bridge method allocs a new web view to be used as a
modal view and the extensions made to web view in SDJSExtensions forces
the web view alloc to happen on the main thread. Because of this we
need to make sure this method is always run on the main thread.